+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - add .idea to .gitignore
 - Add publish\_to: none when localizing
 - Add .gitattributes file
+- Add --git-ref option to localize-refs for custom git refs
 
 ### Changed